backend/requirements.txt
asgiref==3.7.2
axios==0.4.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Django==4.2.3
django-cors-headers==4.2.0
django-urls==1.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Flask==2.3.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
Pillow==10.0.0
Pygments==2.15.1
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
rich==13.5.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
zipp==3.16.2
